moved delete forms out of dashboard
